fix discarding the first 1MB on write.
This was caused by me forgetting that there's already 1mb pending buffer read when starting the file transfer.
this buffer includes the 12 byte packet.

the fix is to undo all the hacky stuff and just add an override to change the max flush size.
this way, we aready properly use whats remaining in the buffer first.